Create is partisipant check

+++ backend/services/message-service.mjs
@@ -5,6 +5,7 @@ import { paginator } from "../utils/paginator.mjs";
 
 export class MessageService {
   #messagesDao = diContainer.resolve(SERVICES.messagesDao);
+  #chatsDao = diContainer.resolve(SERVICES.chatsDao);
 
   async getMessagesByChat(chatId, messagesPerPage, pageNumber) {
     const messagesList = await this.#messagesDao.getMessagesByChat(chatId);
@@ -12,18 +13,28 @@ export class MessageService {
   }
 
   async addMessage(authorId, chatId, messageBody) {
+    const isPatrtisipant = await this.isPatrtisipant(authorId);
+    if (!isPatrtisipant) throw new Error(403);
+
     const messageId = uuidv4();
     const createDate = new Date();
-    await this.#messagesDao.addMessage(chatId, {
+
+    const result = await this.#messagesDao.addMessage(chatId, {
       authorId,
       messageId,
       createDate,
       messageBody,
     });
-    return messageId;
+    if (!result) {
+      throw new Error(500);
+    }
+    return result;
   }
 
   async updateMessage(chatId, messageId, messageBody) {
+    const isPatrtisipant = await this.isPatrtisipant(authorId);
+    if (!isPatrtisipant) throw new Error(403);
+
     const result = await this.#messagesDao.updateMessage(
       chatId,
       messageId,
@@ -37,9 +48,23 @@ export class MessageService {
 
   async deleteMessage(chatId, messageId) {
     try {
+      const isPatrtisipant = await this.isPatrtisipant(authorId);
+      if (!isPatrtisipant) throw new Error(403);
       await this.#messagesDao.deleteMessage(chatId, messageId);
     } catch {
       throw new Error(500);
     }
+  }
+
+  async isPatrtisipant(authorId) {
+    const chats = await this.#chatsDao.getChatsByAuthor(authorId);
+    if (!chats) return false;
+
+    const isPatrtisipant = chats.find((chat) =>
+      chat.participantsIds.includes(authorId)
+    );
+    if (!isPatrtisipant) return false;
+
+    return true;
   }
 }
